% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caagro-package.R
\docType{package}
\name{caagro}
\alias{caagro}
\alias{caagro-package}
\title{caagro: Ciência Aplicada em Análises Agro}
\description{
O pacote \strong{caagro} fornece funções e aplicativos Shiny para
análises estatísticas aplicadas ao setor agrícola.
}
\details{
Funções principais:
\itemize{
\item \code{\link{anQuali}} – Análises de fatores qualitativos.
\item \code{\link{anQuant}} – Análises de fatores quantitativos.
}
}
\author{
\strong{Maintainer}: Wilhan Valasco dos Santos \email{wilhanvalasco@hotmail.com}

}
